chore: refactoring createProjectRevision mutation
