+++ app/schema/schema.graphql
@@ -41367,6 +41367,7 @@ input CreateProjectRevisionInput {
   """
   clientMutationId: String
   projectId: Int!
+  revisionType: String
 }
 
 """The output of our `createProjectRevision` mutation."""
